requirements.txt
allure-pytest==2.14.2
allure-python-commons==2.14.2
alabaster==1.0.0
babel==2.17.0
Brotli==1.1.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.2.0
colorama==0.4.6
docutils==0.21.2
greenlet==3.2.0
idna==3.10
imagesize==1.4.1
iniconfig==2.1.0
Jinja2==3.1.6
markdown2==2.5.3
MarkupSafe==3.0.2
minijinja==2.10.2
packaging==25.0
page==0.3
pexpect==4.9.0
playwright==1.51.0
pluggy==1.5.0
ptyprocess==0.7.0
pyee==12.1.1
Pygments==2.19.1
pytest==8.3.5
pytest-base-url==2.1.0
pytest-playwright==0.7.0
pytest-spec==4.0.0
python-slugify==8.0.4
PyYAML==6.0.2
requests==2.32.3
roman-numerals-py==3.1.0
snowballstemmer==3.0.1
Sphinx==8.2.3
sphinxcontrib-applehelp==2.0.0
sphinxcontrib-devhelp==2.0.0
sphinxcontrib-htmlhelp==2.1.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==2.0.0
sphinxcontrib-serializinghtml==2.0.0
text-unidecode==1.3
typing_extensions==4.13.2
urllib3==2.4.0
pyfunctional==1.5.0
screeninfo==0.8.1